Move onEdit and onDelete calls to parent component
