+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/site/public/locales/en.json
@@ -1056,14 +1056,10 @@
   "Scopes.Usage.UsageViewAPI.api.usage": "List of APIs",
   "ServiceCatalog.Listing.Delete.cancel": "Cancel",
   "ServiceCatalog.Listing.Delete.confirm": "Confirm Delete",
-  "ServiceCatalog.Listing.Delete.error.delete": "Error while deleting service",
   "ServiceCatalog.Listing.Delete.ok.confirm": "Are you sure you want to delete the service {service} ?",
   "ServiceCatalog.Listing.Delete.ok.yes": "Yes",
-  "ServiceCatalog.Listing.Delete.service.deleted.successfully": "Service deleted successfully!",
-  "ServiceCatalog.Listing.Delete.service.updated.successfully": "Service updated successfully!",
   "ServiceCatalog.Listing.Edit.cancel.btn": "Cancel",
   "ServiceCatalog.Listing.Edit.error.retrieve.service.schema.types": "Error while retrieving service and schema types",
-  "ServiceCatalog.Listing.Edit.error.update": "Error while updating service",
   "ServiceCatalog.Listing.Edit.schema.type.label": "Schema Type",
   "ServiceCatalog.Listing.Edit.schema.type.text": "Schema Type of the Service",
   "ServiceCatalog.Listing.Edit.service": "Edit Service",
@@ -1088,13 +1084,17 @@
   "ServiceCatalog.Listing.Listing.create.api": "Create API",
   "ServiceCatalog.Listing.Listing.description1": "From creating and publishing an API to securing, rate-limiting, addresses all aspects of API Management.",
   "ServiceCatalog.Listing.Listing.description2": "From creating and publishing an API to securing, rate-limiting, addresses all aspects of API Management.",
+  "ServiceCatalog.Listing.Listing.error.delete": "Error while deleting service",
   "ServiceCatalog.Listing.Listing.error.loading": "Error While Loading Services",
+  "ServiceCatalog.Listing.Listing.error.update": "Error while updating service",
   "ServiceCatalog.Listing.Listing.get.started": "Get Started",
   "ServiceCatalog.Listing.Listing.heading": "Service Catalog",
   "ServiceCatalog.Listing.Listing.help.tooltip": "The Service Catalog enables API-first Integration",
   "ServiceCatalog.Listing.Listing.name": "Service",
   "ServiceCatalog.Listing.Listing.schema.type": "Schema Type",
+  "ServiceCatalog.Listing.Listing.service.deleted.successfully": "Service deleted successfully!",
   "ServiceCatalog.Listing.Listing.service.type": "Service Type",
+  "ServiceCatalog.Listing.Listing.service.updated.successfully": "Service updated successfully!",
   "ServiceCatalog.Listing.Listing.service.url": "Service URL",
   "ServiceCatalog.Listing.Listing.usage": "No. Of APIs",
   "ServiceCatalog.Listing.Listing.version": "Version",
