+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher/site/public/locales/fr.json
@@ -1056,14 +1056,10 @@
   "Scopes.Usage.UsageViewAPI.api.usage": "",
   "ServiceCatalog.Listing.Delete.cancel": "",
   "ServiceCatalog.Listing.Delete.confirm": "",
-  "ServiceCatalog.Listing.Delete.error.delete": "",
   "ServiceCatalog.Listing.Delete.ok.confirm": "",
   "ServiceCatalog.Listing.Delete.ok.yes": "",
-  "ServiceCatalog.Listing.Delete.service.deleted.successfully": "",
-  "ServiceCatalog.Listing.Delete.service.updated.successfully": "",
   "ServiceCatalog.Listing.Edit.cancel.btn": "",
   "ServiceCatalog.Listing.Edit.error.retrieve.service.schema.types": "",
-  "ServiceCatalog.Listing.Edit.error.update": "",
   "ServiceCatalog.Listing.Edit.schema.type.label": "",
   "ServiceCatalog.Listing.Edit.schema.type.text": "",
   "ServiceCatalog.Listing.Edit.service": "",
@@ -1088,13 +1084,17 @@
   "ServiceCatalog.Listing.Listing.create.api": "",
   "ServiceCatalog.Listing.Listing.description1": "",
   "ServiceCatalog.Listing.Listing.description2": "",
+  "ServiceCatalog.Listing.Listing.error.delete": "",
   "ServiceCatalog.Listing.Listing.error.loading": "",
+  "ServiceCatalog.Listing.Listing.error.update": "",
   "ServiceCatalog.Listing.Listing.get.started": "",
   "ServiceCatalog.Listing.Listing.heading": "",
   "ServiceCatalog.Listing.Listing.help.tooltip": "",
   "ServiceCatalog.Listing.Listing.name": "",
   "ServiceCatalog.Listing.Listing.schema.type": "",
+  "ServiceCatalog.Listing.Listing.service.deleted.successfully": "",
   "ServiceCatalog.Listing.Listing.service.type": "",
+  "ServiceCatalog.Listing.Listing.service.updated.successfully": "",
   "ServiceCatalog.Listing.Listing.service.url": "",
   "ServiceCatalog.Listing.Listing.usage": "",
   "ServiceCatalog.Listing.Listing.version": "",
